align data --TODO: align data in config

+++ include/EngineMain.h
@@ -15,7 +15,7 @@ namespace Engine
 
 		public:
 		OperationArchitecture::SystemBus *SystemBus;
-		EngineMain(const void *config, unsigned int &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
+		EngineMain(const void *config, size_t &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
 
 		void Setup();
 		void Loop();
