+++ include/Operations/Operation_CylinderAirMass_SD.h
@@ -13,7 +13,7 @@ namespace OperationArchitecture
 
 		float Execute(float cylinderAirTemperature, float manifoldAbsolutePressure, float VolumetricEfficiency) override;
 		
-		static IOperationBase *Create(const void *config, unsigned int &sizeOut);
+		static IOperationBase *Create(const void *config, size_t &sizeOut);
 	};
 }
 #endif
