+++ include/Operations/Operation_EngineInjectorPrime.h
@@ -18,7 +18,7 @@ namespace OperationArchitecture
 
 		void Execute(float time) override;
 
-		static IOperationBase *Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection, OperationPackager *packager);
+		static IOperationBase *Create(const void *config, size_t &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection, OperationPackager *packager);
 	};
 }
 #endif
