+++ include/Operations/Operation_EnginePositionPrediction.h
@@ -15,7 +15,7 @@ namespace OperationArchitecture
 
 		uint32_t Execute(float desiredPosition, EnginePosition enginePosition) override;
 
-		static IOperationBase *Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
+		static IOperationBase *Create(const void *config, size_t &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
 	};
 }
 #endif
