+++ include/Operations/Operation_EngineScheduleIgnition.h
@@ -28,7 +28,7 @@ namespace OperationArchitecture
 		void Dwell();
 		void Ignite();
 
-		static IOperationBase *Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection, OperationPackager *packager);
+		static IOperationBase *Create(const void *config, size_t &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection, OperationPackager *packager);
 	};
 }
 #endif
