+++ src/Operations/Operation_EngineInjectorPrime.cpp
@@ -18,12 +18,12 @@ namespace OperationArchitecture
 		_timerService->ScheduleTask(_closeTask, _timerService->GetTick() + time * _timerService->GetTicksPerSecond());
 	}
 
-	IOperationBase *Operation_EngineInjectorPrime::Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection, OperationPackager *packager)
+	IOperationBase *Operation_EngineInjectorPrime::Create(const void *config, size_t &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection, OperationPackager *packager)
 	{
 		ICallBack * openCallBack = 0;
 		ICallBack * closeCallBack = 0;
 
-		unsigned int size = 0;
+		size_t size = 0;
 		IOperationBase *operation = packager->Package(config, size);
 		Config::OffsetConfig(config, sizeOut, size);
 		if(operation->NumberOfParameters == 1)
